[ETC] 계단 / - / 1시간 폴더 구조 변경 및 클래스 분리
