+++ shkisme/ETC/stair/Position.java
@@ -1,0 +1,5 @@
+package stair;
+
+public enum Position {
+  ON, IN, OUT
+}
